feat: refactored Checkbox component to use react-hook-form and improve structure
