+++ src/atomic/atm.checkbox/index.ts
@@ -1,0 +1,3 @@
+import Checkbox from './checkbox.component';
+
+export { Checkbox };
